organisation de l'architecture pour le merge et rectification du profil-test
