+++ backend/src/main/java/com/example/backend/config/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package com.example.backend.configswagger;
+package com.example.backend.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Contact;
