+++ backend/src/main/java/com/example/backend/controllers/AuthController.java
@@ -1,9 +1,9 @@
-package com.example.backend.auth.authController;
+package com.example.backend.controllers;
 
-import com.example.backend.auth.authController.authservice.AuthService;
-import com.example.backend.entityDTO.authdto.AuthResponseDTO;
-import com.example.backend.entityDTO.authdto.LoginDTO;
-import com.example.backend.entityDTO.authdto.RegisterDTO;
+import com.example.backend.services.AuthService;
+import com.example.backend.dto.authdto.AuthResponseDTO;
+import com.example.backend.dto.authdto.LoginDTO;
+import com.example.backend.dto.authdto.RegisterDTO;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.*;
