+++ backend/src/main/java/com/example/backend/controllers/reservationcontroller/ReservationController.java
@@ -1,9 +1,9 @@
-package com.example.backend.auth.authController.reservationcontroller;
+package com.example.backend.controllers.reservationcontroller;
 
-import com.example.backend.entityDTO.reservationdto.DisponibiliteRequestDTO;
-import com.example.backend.entityDTO.reservationdto.ReservationRequestDTO;
-import com.example.backend.entityDTO.reservationdto.ReservationResponseDTO;
-import com.example.backend.reservationservice.ReservationService;
+import com.example.backend.dto.reservationdto.DisponibiliteRequestDTO;
+import com.example.backend.dto.reservationdto.ReservationRequestDTO;
+import com.example.backend.dto.reservationdto.ReservationResponseDTO;
+import com.example.backend.services.reservationservice.ReservationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
