+++ backend/src/main/java/com/example/backend/dto/authdto/AuthResponseDTO.java
@@ -1,4 +1,4 @@
-package com.example.backend.entityDTO.authdto;
+package com.example.backend.dto.authdto;
 
 import com.example.backend.roles.RoleUtilisateur;
