+++ backend/src/main/java/com/example/backend/entities/Admin.java
@@ -11,7 +11,7 @@ import java.util.List;
  * @author Wulfrid MBONGO
  */
 @Entity
-@Table(name = "admin", catalog = "logement_cameroun", schema = "")
+@Table(name = "admin", catalog = "logement_cameroun")
 @NamedQueries({
     @NamedQuery(name = "Admin.findAll", query = "SELECT a FROM Admin a"),
     @NamedQuery(name = "Admin.findById", query = "SELECT a FROM Admin a WHERE a.id = :id"),
