+++ backend/src/main/java/com/example/backend/entities/Disponibilite.java
@@ -11,7 +11,7 @@ import java.util.Date;
  * @author Wulfrid MBONGO
  */
 @Entity
-@Table(name = "disponibilite", catalog = "logement_cameroun", schema = "")
+@Table(name = "disponibilite", catalog = "logement_cameroun")
 @NamedQueries({
     @NamedQuery(name = "Disponibilite.findAll", query = "SELECT d FROM Disponibilite d"),
     @NamedQuery(name = "Disponibilite.findById", query = "SELECT d FROM Disponibilite d WHERE d.id = :id"),
