+++ backend/src/main/java/com/example/backend/entities/Equipement.java
@@ -10,7 +10,7 @@ import java.util.List;
  * @author Wulfrid MBONGO
  */
 @Entity
-@Table(name = "equipement", catalog = "logement_cameroun", schema = "")
+@Table(name = "equipement", catalog = "logement_cameroun")
 @NamedQueries({
     @NamedQuery(name = "Equipement.findAll", query = "SELECT e FROM Equipement e"),
     @NamedQuery(name = "Equipement.findById", query = "SELECT e FROM Equipement e WHERE e.id = :id"),
