+++ backend/src/main/java/com/example/backend/entities/Localisation.java
@@ -9,7 +9,7 @@ import java.util.List;
  * @author Wulfrid MBONGO
  */
 @Entity
-@Table(name = "localisation", catalog = "logement_cameroun", schema = "")
+@Table(name = "localisation", catalog = "logement_cameroun")
 @NamedQueries({
     @NamedQuery(name = "Localisation.findAll", query = "SELECT l FROM Localisation l"),
     @NamedQuery(name = "Localisation.findById", query = "SELECT l FROM Localisation l WHERE l.id = :id"),
