+++ backend/src/main/java/com/example/backend/entities/Paiement.java
@@ -10,7 +10,7 @@ import java.util.List;
  * @author Wulfrid MBONGO
  */
 @Entity
-@Table(name = "paiement", catalog = "logement_cameroun", schema = "")
+@Table(name = "paiement", catalog = "logement_cameroun")
 @NamedQueries({
     @NamedQuery(name = "Paiement.findAll", query = "SELECT p FROM Paiement p"),
     @NamedQuery(name = "Paiement.findById", query = "SELECT p FROM Paiement p WHERE p.id = :id"),
