+++ backend/src/main/java/com/example/backend/entities/Reservation.java
@@ -12,7 +12,7 @@ import java.util.List;
  * @author Wulfrid MBONGO
  */
 @Entity
-@Table(name = "reservation", catalog = "logement_cameroun", schema = "")
+@Table(name = "reservation", catalog = "logement_cameroun")
 @NamedQueries({
     @NamedQuery(name = "Reservation.findAll", query = "SELECT r FROM Reservation r"),
     @NamedQuery(name = "Reservation.findById", query = "SELECT r FROM Reservation r WHERE r.id = :id"),
