+++ backend/src/main/java/com/example/backend/entities/Utilisateur.java
@@ -10,7 +10,7 @@ import java.util.List;
  * @author Wulfrid MBONGO
  */
 @Entity
-@Table(name = "utilisateur", catalog = "logement_cameroun", schema = "")
+@Table(name = "utilisateur", catalog = "logement_cameroun")
 @NamedQueries({
     @NamedQuery(name = "Utilisateur.findAll", query = "SELECT u FROM Utilisateur u"),
     @NamedQuery(name = "Utilisateur.findById", query = "SELECT u FROM Utilisateur u WHERE u.id = :id"),
