+++ backend/src/main/java/com/example/backend/entities/Voyageur.java
@@ -10,7 +10,7 @@ import java.util.List;
  * @author Wulfrid MBONGO
  */
 @Entity
-@Table(name = "voyageur", catalog = "logement_cameroun", schema = "")
+@Table(name = "voyageur", catalog = "logement_cameroun")
 @NamedQueries({
     @NamedQuery(name = "Voyageur.findAll", query = "SELECT v FROM Voyageur v"),
     @NamedQuery(name = "Voyageur.findById", query = "SELECT v FROM Voyageur v WHERE v.id = :id")})
