+++ backend/src/main/resources/application-test.properties
@@ -1,16 +1,18 @@
+# Configuration H2 pour les tests
 spring.datasource.driver-class-name=org.h2.Driver
 spring.datasource.url=jdbc:h2:mem:testdb;MODE=MySQL;DATABASE_TO_LOWER=TRUE;CASE_INSENSITIVE_IDENTIFIERS=TRUE
 spring.datasource.username=sa
 spring.datasource.password=
 
+# Configuration JPA/Hibernate
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.show-sql=false
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 
-# IMPORTANT : Ignorer le catalog dans les tests
+# Ignorer le catalog en mode test (IMPORTANT)
 spring.jpa.properties.hibernate.default_catalog=
 spring.jpa.properties.hibernate.globally_quoted_identifiers=false
-spring.jpa.properties.hibernate.globally_quoted_identifiers_skip_column_definitions=true
 
+# Configuration JWT
 jwt.secret=testSecretKeyForJwtThatIsAtLeast256BitsLongForHS256Algorithm
 jwt.expiration=86400000
