+++ backend/src/main/resources/application.properties
@@ -1,2 +1,2 @@
 spring.application.name=backend
-spring.properties.active=test
+spring.profiles.active=test
