+++ backend/src/test/java/com/example/backend/auth/authController/authservice/AuthServiceTest.java
@@ -1,12 +1,13 @@
 package com.example.backend.auth.authController.authservice;
 
 import com.example.backend.entities.*;
-import com.example.backend.entityDTO.authdto.AuthResponseDTO;
-import com.example.backend.entityDTO.authdto.LoginDTO;
-import com.example.backend.entityDTO.authdto.RegisterDTO;
+import com.example.backend.dto.authdto.AuthResponseDTO;
+import com.example.backend.dto.authdto.LoginDTO;
+import com.example.backend.dto.authdto.RegisterDTO;
 import com.example.backend.repository.*;
 import com.example.backend.roles.RoleUtilisateur;
 import com.example.backend.security.jwt.JwtUtils;
+import com.example.backend.services.AuthService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
