+++ backend/src/test/java/com/example/backend/reservationservice/ReservationServiceTest.java
@@ -4,13 +4,14 @@ import com.example.backend.entities.Annonces;
 import com.example.backend.entities.Disponibilite;
 import com.example.backend.entities.Reservation;
 import com.example.backend.entities.Voyageur;
-import com.example.backend.entityDTO.reservationdto.DisponibiliteRequestDTO;
-import com.example.backend.entityDTO.reservationdto.ReservationRequestDTO;
-import com.example.backend.entityDTO.reservationdto.ReservationResponseDTO;
+import com.example.backend.dto.reservationdto.DisponibiliteRequestDTO;
+import com.example.backend.dto.reservationdto.ReservationRequestDTO;
+import com.example.backend.dto.reservationdto.ReservationResponseDTO;
 import com.example.backend.repository.AnnoncesRepository;
 import com.example.backend.repository.DisponibiliteRepository;
 import com.example.backend.repository.ReservationRepository;
 import com.example.backend.repository.VoyageurRepository;
+import com.example.backend.services.reservationservice.ReservationService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
